feat: added support to own tokens
