+++ src/components/TokensModal/TokensModal.tsx
@@ -1,48 +1,11 @@
 import Image from "next/image";
 import FullscreenIcon from "public/fullscreen.svg";
 import { useState } from "react";
-import { COLORS, ICON_SIZES, truncateString } from "~/utils";
-import { TOKENS, tokensData, type Asset } from "~/utils/tokens";
+import { COLORS, ICON_SIZES } from "~/utils";
+import { tokensData, type Asset } from "~/utils/tokens";
 import { Button, SIZE } from "../Button/Button";
 import { Modal } from "../Modal/Modal";
-
-interface TokenElementProps {
-  token: [string, Asset];
-}
-
-const TokenElement = ({ token }: TokenElementProps) => {
-  const [hiddenPolicy, setHiddenPolicy] = useState(true);
-  const changeVisibility = () => setHiddenPolicy(!hiddenPolicy);
-
-  return (
-    <div className="flex items-center justify-between">
-      <div className="flex items-center gap-3">
-        {token[1].icon}
-        <span className="text-sm font-medium">
-          {truncateString(token[0], 7)}
-        </span>
-        {token[0] !== TOKENS.ADA && (
-          <span
-            className="cursor-pointer text-xs font-medium"
-            onClick={changeVisibility}
-          >
-            {hiddenPolicy
-              ? truncateString(token[1].policyId, 10)
-              : token[1].policyId}
-          </span>
-        )}
-      </div>
-      <div className="flex items-center gap-4">
-        <span className="text-sm font-medium">
-          Decimals: {token[1].decimals}
-        </span>
-        <div className="w-fit">
-          <Button size={SIZE.XSMALL}>Select</Button>
-        </div>
-      </div>
-    </div>
-  );
-};
+import { TokenElement } from "./TokenElement";
 
 interface TokensModalProps {
   assets?: [string, Asset];
